refactor: separate KeychainItem into new file
#10
