+++ PhotoTag/PhotoTag/Login/Model/KeychainItem.swift
@@ -1,0 +1,123 @@
+//
+//  KeychainItem.swift
+//  PhotoTag
+//
+//  Created by Keunna Lee on 2020/11/04.
+//
+
+import Foundation
+
+struct KeychainItem {
+    
+    // MARK: Types
+    enum KeychainError: Error {
+        case noPassword
+        case unexpectedPasswordData
+        case unexpectedItemData
+        case unhandledError
+    }
+    
+    // MARK: Properties
+    let service: String
+    private(set) var account: String
+    let accessGroup: String?
+    
+    // MARK: Intialization
+    init(service: String, account: String, accessGroup: String? = nil) {
+        self.service = service
+        self.account = account
+        self.accessGroup = accessGroup
+    }
+    
+    // MARK: Keychain access
+    func readItem() throws -> String {
+        var query = KeychainItem.keychainQuery(withService: service, account: account, accessGroup: accessGroup)
+        query[kSecMatchLimit as String] = kSecMatchLimitOne
+        query[kSecReturnAttributes as String] = kCFBooleanTrue
+        query[kSecReturnData as String] = kCFBooleanTrue
+        
+        var queryResult: AnyObject?
+        let status = withUnsafeMutablePointer(to: &queryResult) {
+            SecItemCopyMatching(query as CFDictionary, UnsafeMutablePointer($0))
+        }
+        
+        guard status != errSecItemNotFound else { throw KeychainError.noPassword }
+        guard status == noErr else { throw KeychainError.unhandledError }
+        
+        guard let existingItem = queryResult as? [String: AnyObject],
+            let passwordData = existingItem[kSecValueData as String] as? Data,
+            let password = String(data: passwordData, encoding: String.Encoding.utf8)
+            else {
+                throw KeychainError.unexpectedPasswordData
+        }
+        
+        return password
+    }
+    
+    func saveItem(_ password: String) throws {
+        let encodedPassword = password.data(using: String.Encoding.utf8)!
+        
+        do {
+            try _ = readItem()
+            
+            var attributesToUpdate = [String: AnyObject]()
+            attributesToUpdate[kSecValueData as String] = encodedPassword as AnyObject?
+            
+            let query = KeychainItem.keychainQuery(withService: service, account: account, accessGroup: accessGroup)
+            let status = SecItemUpdate(query as CFDictionary, attributesToUpdate as CFDictionary)
+            
+            guard status == noErr else { throw KeychainError.unhandledError }
+        } catch KeychainError.noPassword {
+            var newItem = KeychainItem.keychainQuery(withService: service, account: account, accessGroup: accessGroup)
+            newItem[kSecValueData as String] = encodedPassword as AnyObject?
+            
+            let status = SecItemAdd(newItem as CFDictionary, nil)
+            
+            guard status == noErr else { throw KeychainError.unhandledError }
+        }
+    }
+    
+    func deleteItem() throws {
+        // Delete the existing item from the keychain.
+        let query = KeychainItem.keychainQuery(withService: service, account: account, accessGroup: accessGroup)
+        let status = SecItemDelete(query as CFDictionary)
+        
+        // Throw an error if an unexpected status was returned.
+        guard status == noErr || status == errSecItemNotFound else { throw KeychainError.unhandledError }
+    }
+    
+    // MARK: Convenience
+    
+    private static func keychainQuery(withService service: String, account: String? = nil, accessGroup: String? = nil) -> [String: AnyObject] {
+        var query = [String: AnyObject]()
+        query[kSecClass as String] = kSecClassGenericPassword
+        query[kSecAttrService as String] = service as AnyObject?
+        
+        if let account = account {
+            query[kSecAttrAccount as String] = account as AnyObject?
+        }
+        
+        if let accessGroup = accessGroup {
+            query[kSecAttrAccessGroup as String] = accessGroup as AnyObject?
+        }
+        
+        return query
+    }
+    
+    static var currentUserIdentifier: String {
+        do {
+            let storedIdentifier = try KeychainItem(service: "lena.SimLeeTag.PhotoTag", account: "userIdentifier").readItem()
+            return storedIdentifier
+        } catch {
+            return ""
+        }
+    }
+    
+    static func deleteUserIdentifierFromKeychain() {
+        do {
+            try KeychainItem(service: "lena.SimLeeTag.PhotoTag", account: "userIdentifier").deleteItem()
+        } catch {
+            print("Unable to delete userIdentifier from keychain")
+        }
+    }
+}
